+++ companion/CMakeLists.txt
@@ -31,7 +31,7 @@ set(DEV_BUILD false CACHE STRING "If true, the application is built for developm
 message("DEV_BUILD: ${DEV_BUILD}")
 
 if(UNIX)
-    list(APPEND SOURCE_LIST "${CMAKE_CURRENT_SOURCE_DIR}/src/logutil.cpp")
+    list(APPEND SOURCE_LIST "${CMAKE_CURRENT_SOURCE_DIR}/src/logutil_linux.cpp")
     list(APPEND SOURCE_LIST "${CMAKE_CURRENT_SOURCE_DIR}/src/windowhandler_linux.cpp")
 endif()
